+++ src/main/java/com/he/attend/controller/RoleController.java
@@ -4,17 +4,26 @@ package com.he.attend.controller;
 import com.baomidou.mybatisplus.mapper.EntityWrapper;
 import com.he.attend.common.JsonResult;
 import com.he.attend.common.PageResult;
+import com.he.attend.common.utils.PowerUtil;
 import com.he.attend.model.Role;
+import com.he.attend.model.RoleAuthorities;
+import com.he.attend.model.UserRole;
+import com.he.attend.service.RoleAuthoritiesService;
 import com.he.attend.service.RoleService;
+import com.he.attend.service.UserRoleService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import org.wf.jwtp.provider.Token;
+import org.wf.jwtp.util.SubjectUtil;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.Iterator;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Api(value = "角色相关的接口", tags = "role")
 @RestController
@@ -22,14 +31,17 @@ import java.util.List;
 public class RoleController {
     @Autowired
     private RoleService roleService;
-
+    @Autowired
+    private UserRoleService userRoleService;
+    @Autowired
+    private RoleAuthoritiesService roleAuthoritiesService;
     @ApiOperation(value = "查询所有角色")
     @ApiImplicitParams({
             @ApiImplicitParam(name = "keyword", value = "搜索关键字", required = false, dataType = "String", paramType = "query"),
             @ApiImplicitParam(name = "access_token", value = "令牌", required = true, dataType = "String", paramType = "query")
     })
     @GetMapping()
-    public PageResult<Role> list(String keyword) {
+    public PageResult<Role> list(String keyword,HttpServletRequest request) {
         List<Role> list = roleService.selectList(new EntityWrapper<Role>().orderBy("create_time", true));
         // 筛选结果
         if (keyword != null && !keyword.trim().isEmpty()) {
@@ -52,7 +64,12 @@ public class RoleController {
             @ApiImplicitParam(name = "access_token", value = "令牌", required = true, dataType = "String", paramType = "form")
     })
     @PostMapping()
-    public JsonResult add(Role role) {
+    public JsonResult add(Role role,HttpServletRequest request) {
+        String authority="post:/role";
+        boolean canDo = PowerUtil.powerPermisson(authority,request);
+        if(!canDo){
+            return  JsonResult.error("没有相应权限");
+        }
         if (roleService.insert(role)) {
             return JsonResult.ok("添加成功");
         }
@@ -65,7 +82,12 @@ public class RoleController {
             @ApiImplicitParam(name = "access_token", value = "令牌", required = true, dataType = "String", paramType = "form")
     })
     @PutMapping()
-    public JsonResult update(Role role) {
+    public JsonResult update(Role role,HttpServletRequest request) {
+        String authority="put:/role";
+        boolean canDo = PowerUtil.powerPermisson(authority,request);
+        if(!canDo){
+            return  JsonResult.error("没有相应权限");
+        }
         if (roleService.updateById(role)) {
             return JsonResult.ok("修改成功！");
         }
@@ -78,10 +100,17 @@ public class RoleController {
             @ApiImplicitParam(name = "access_token", value = "令牌", required = true, dataType = "String", paramType = "query")
     })
     @DeleteMapping("/{id}")
-    public JsonResult delete(@PathVariable("id") Integer roleId) {
+    public JsonResult delete(@PathVariable("id") Integer roleId, HttpServletRequest request) {
+        String authority="delete:/role";
+        boolean canDo = PowerUtil.powerPermisson(authority,request);
+        if(!canDo){
+            return  JsonResult.error("没有相应权限");
+        }
         if (roleService.deleteById(roleId)) {
             return JsonResult.ok("删除成功");
         }
         return JsonResult.error("删除失败");
     }
+
+
 }
