+++ src/main/java/com/he/attend/controller/StaffController.java
@@ -1,12 +1,16 @@
 package com.he.attend.controller;
 
+import com.alibaba.fastjson.JSONObject;
 import com.baomidou.mybatisplus.mapper.EntityWrapper;
 import com.baomidou.mybatisplus.plugins.Page;
 import com.he.attend.common.JsonResult;
 import com.he.attend.common.PageResult;
+import com.he.attend.common.utils.PowerUtil;
 import com.he.attend.model.*;
+import com.he.attend.service.AttendService;
 import com.he.attend.service.RuleService;
 import com.he.attend.service.StaffService;
+import com.wangfan.endecrypt.utils.EndecryptUtils;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
@@ -16,6 +20,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -27,12 +32,18 @@ public class StaffController {
 
     @Autowired
     private StaffService staffService;
+    @Autowired
+    private AttendService attendService;
 
     @Autowired
     private RuleService ruleService;
     @RequestMapping("/insert")
-    public PageResult insertStaff(Staff staff){
-
+    public PageResult insertStaff(Staff staff, HttpServletRequest request){
+        String authority="post:/staff/insert";
+        boolean canDo = PowerUtil.powerPermisson(authority,request);
+        if(!canDo){
+            return new PageResult("没有相应的权限",400);
+        }
         if(staffService.insert(staff)){
             return new PageResult("添加成功",200);
         }
@@ -40,8 +51,12 @@ public class StaffController {
     }
 
     @RequestMapping("/update")
-    public PageResult updateStaff(Staff staff){
-
+    public PageResult updateStaff(Staff staff,HttpServletRequest request){
+        String authority="put:/staff/update";
+        boolean canDo = PowerUtil.powerPermisson(authority,request);
+        if(!canDo){
+            return new PageResult("没有相应的权限",400);
+        }
         if(staffService.updateById(staff)){
             return new PageResult("修改成功",200);
         }
@@ -49,8 +64,12 @@ public class StaffController {
     }
 
     @RequestMapping("/delete/{staffId}")
-    public PageResult delteStaff(@PathVariable("staffId") Integer staffId){
-
+    public PageResult delteStaff(@PathVariable("staffId") Integer staffId,HttpServletRequest request){
+        String authority="delete:/staff/delete";
+        boolean canDo = PowerUtil.powerPermisson(authority,request);
+        if(!canDo){
+            return new PageResult("没有相应的权限",400);
+        }
         if(staffService.deleteById(staffId)){
             return new PageResult("删除成功",200);
         }
@@ -94,7 +113,12 @@ public class StaffController {
             @ApiImplicitParam(name = "access_token", value = "令牌", required = true, dataType = "String", paramType = "form")
     })
     @PutMapping("/state")
-    public JsonResult updateState(Integer staffId, Integer state) {
+    public JsonResult updateState(Integer staffId, Integer state,HttpServletRequest request) {
+        String authority="put:/staff/statu";
+        boolean canDo = PowerUtil.powerPermisson(authority,request);
+        if(!canDo){
+            return JsonResult.error("没有相应的权限");
+        }
         if (state == null || (state != 0 && state != 1)) {
             return JsonResult.error("state值需要在[0,1]中");
         }
@@ -107,6 +131,26 @@ public class StaffController {
         return JsonResult.error();
     }
 
+    @ApiOperation(value = "重置密码", notes = "")
+    @ApiImplicitParams({
+            @ApiImplicitParam(name = "userId", value = "用户id", required = true, dataType = "String", paramType = "path"),
+            @ApiImplicitParam(name = "access_token", value = "令牌", required = true, dataType = "String", paramType = "form")
+    })
+    @PutMapping("/psw/{id}")
+    public JsonResult resetPsw(@PathVariable("id") Integer staffId,HttpServletRequest request) {
+        String authority="put:/staff/psw/{id}";
+        boolean canDo = PowerUtil.powerPermisson(authority,request);
+        if(!canDo){
+            return JsonResult.error("没有相应的权限");
+        }
+        Staff staff = new Staff();
+        staff.setStaffId(staffId);
+        staff.setPassword(("123456"));
+        if (staffService.updateById(staff)) {
+            return JsonResult.ok("重置密码成功");
+        }
+        return JsonResult.error("重置密码失败");
+    }
     /**
      * 查询所有状态为0 的员工
      * @return
@@ -148,5 +192,69 @@ public class StaffController {
             }
         }
         return new PageResult<>(200,"查询员工成功",finalStaff.size(),finalStaff);
+    }
+
+
+    /**
+     * 查询今日打卡和未打卡人数
+     */
+    @RequestMapping("/queryAttendAndNotAttend")
+    public JSONObject queryAttendAndNotAttend(){
+        //所有员工人数
+        Integer staffNum=staffService.queryStaffNum();
+        List<Attend> todayAttend=attendService.queryTodayAttend();
+        List<Integer> havaAttendStaffIds=new ArrayList<>();//已打卡员工的id
+        for(Attend attend:todayAttend){
+            if(!havaAttendStaffIds.contains(attend.getStaffId())){
+                havaAttendStaffIds.add(attend.getStaffId());
+            }
+        }
+        JSONObject jsonObject=new JSONObject();
+        jsonObject.put("attendnum",havaAttendStaffIds.size());
+        jsonObject.put("notattendnum",staffNum-havaAttendStaffIds.size());
+        return jsonObject;
+    }
+
+    /**
+     * 查询今日打卡各个类型的人数
+     */
+    @RequestMapping("/queryAttendTypeNum")
+    public JSONObject queryAttendTypeNum(){
+        //所有员工人数
+        Integer staffNum=staffService.queryStaffNum();
+        List<Attend> todayAttend=attendService.queryTodayAttend();
+        List<Integer> havaAttendStaffIds=new ArrayList<>();//已打卡员工的id
+        Integer normal=0;
+        Integer lateAttendNum=0;
+        Integer earlyAttendNum=0;
+
+        Integer inAttend=0;//内勤打卡人数
+        Integer outAttend=0;//外勤打卡人数
+        for(Attend attend:todayAttend){
+            if(!havaAttendStaffIds.contains(attend.getStaffId())){
+                havaAttendStaffIds.add(attend.getStaffId());
+            }
+            if(attend.getType().equals("正常")){
+                normal+=1;
+            }else if(attend.getType().equals("迟到")){
+                lateAttendNum+=1;
+            }else if(attend.getType().equals("早退")){
+                earlyAttendNum+=1;
+            }
+
+            if(attend.getPlaceType().equals("内勤")){
+                inAttend+=1;
+            }else {
+                outAttend+=1;
+            }
+        }
+        JSONObject jsonObject=new JSONObject();
+        jsonObject.put("normal",normal);
+        jsonObject.put("lateAttendNum",lateAttendNum);
+        jsonObject.put("earlyAttendNum",earlyAttendNum);
+        jsonObject.put("inAttend",inAttend);
+        jsonObject.put("outAttend",outAttend);
+        jsonObject.put("notattendnum",staffNum-havaAttendStaffIds.size());
+        return jsonObject;
     }
 }
